Also binary sensor is now creted
